app/requirements.txt
streamlit
ollama
openai
requests
Pillow
PyJWT
cryptography
extra_streamlit_components
toml
streamlit-authenticator
pytz

pdf2image
poppler-utils
langchain
langchain-core
langchain-community
langchain_text_splitters
unstructured
unstructured[all-docs]
pdfminer.six